Add files via upload
Manipulate, clean, and analyze data from a randomized evaluation. A randomized control trial with 432 small and medium enterprises in Mexico shows positive impact of access to 1
year of management consulting services on total factor productivity and return on assets. Owners also had an increase
in “entrepreneurial spirit” (an index that measures entrepreneurial confidence and goal setting). Using Mexican social
security data, we find a persistent large increase (about 50 percent) in the number of employees and total wage bill
even 5 years after the program. We document large heterogeneity in the specific managerial practices that improved
as a result of the consulting, with the most prominent being marketing, financial accounting, and long-term business
planning.